解决加密文件由于文件指针类型不一致，导致的seek异常。

+++ DbstarDVB/jni/dvbpush/include/drmapi.h
@@ -8,7 +8,7 @@ int drm_set_emmpid();
 void drm_uninit();
 int drm_open(int fd1, int fd2);
 int drm_read(int fd, unsigned char *buf, int size);
-int64_t drm_seek(int fd, int pos, int whence);
+int64_t drm_seek(int fd, int64_t pos, int whence);
 void drm_close(int fd);
 
 #endif
